refactor: progress: split VecProgress into separate modules
Extract `VecProgress` implementation and its dependencies from
`progress/mod.rs` into separate module files to improve code
organization and modularity. The progress module is now structured
with focused, single-responsibility modules.
